+++ extensions/positron-r/package.nls.json
@@ -3,6 +3,10 @@
 	"description": "R Language Pack for Positron",
 	"r.command.setKernelPath.title": "Set Kernel Path",
 	"r.command.createNewFile.title": "New R File",
+	"r.command.packageLoad.title": "Load R Package",
+	"r.command.packageBuild.title": "Build R Package",
+	"r.command.packageTest.title": "Test R Package",
+	"r.command.packageCheck.title": "Check R Package",
 	"r.menu.createNewFile.title": "R File",
 	"r.configuration.title": "Positron R Language Pack",
 	"r.configuration.kernelPath.description": "Path on disk to the ARK kernel executable; use this to override the default (embedded) kernel.",
